requirements.txt
qrcode
pycryptodome